airflow/requirements.txt
Apache Airflow==2.8.0
pydantic==2.5.0
pymongo==4.6.0
requests==2.31.0
python-dotenv==1.0.0
cryptography==41.0.7
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
